bucketpermissions/requirements.txt
google-cloud-storage==1.2.0
google-cloud-resource-manager==0.25.0
httplib2==0.10.3
parse==1.8.2
python-dateutil==2.6.1